Refactor user-facing API to separate subpackage (#101)
* Refactor to include user-facing API

* Ignore missing type in empty array.

* Include api methods with api.__all__, del api in __init__.py.
