+++ meeshkan/api/__init__.py
@@ -1,0 +1,12 @@
+"""User-facing API is implemented in this submodule!"""
+
+from . import conditions
+from .conditions import *  # pylint: disable=wildcard-import
+from . import scalars
+from .scalars import *  # pylint: disable=wildcard-import
+from . import utils
+from .utils import *  # pylint: disable=wildcard-import
+
+
+__all__ = scalars.__all__
+__all__ += conditions.__all__
